requirements.txt
anthropic
streamlit
python-dotenv
youtube-transcript-api
python-docx
requests
pyyaml
PyPDF2
